Remove vpp install-dep from build script
This saves time, and also can hang if there are some updates to be
installed. The only drawback is the need to update the Dockerfile when
the VPP dependencies change, which is very infrequent - and only happens
when we change the VPP commit we use, meaning we need to rebuild an
image to test it, and have to notice that a change is needed.