agent: clean conflicting IPs on pod add
It can happen that having already configured an address
for a pod in VPP, we receive an add for a new pod with this
same address. The deletion notification for the old pod
comes only later, removing the main VRF route for the new
pod.

This patch ensures that when we adding the second pod we
first make sure to clean up existing pods with this address.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>