agent: deduce mtu from felix+ipam & update in pods
This patch deduces the default MTU to use in pods
from the felix configuration, taking into account
updates that were not well propagated to the cni_server
object. It also takes into account the ippools present
when they have ipip or vxlan enabled, as this is not
reflected in the felixconfiguration.

On updates, the pods are also walked and their interface MTUs
are updated.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>