vpp-manager: Fix corefile cleanup
This patch fixes a segfault happening when the corefile was not specified
in the environment variable. It also tries to improve the logic a bit
and adds a test for this function.

Additionnally this patch also enables running tests in `make test` and in
the CI. This should not include the integration tests that require the
envvar INTEGRATION_TEST to be set.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>