vpp-manager: if annotation only on uplinks
This patch moves the interface annotations to the uplink interface
only as they are now only used here. They were initally introduced
on all kind of interfaces for future proofing, but as `struc` does
not allow maps out of the box to serialize data on disk we'll
defer to future usecase to re-introduce annotation on pod interfaces.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>