vpp: use capo RX/TX in vpp & ingress/egress in Calico.
This patch changes the semantics for capo policies and rules,
it makes it so that VPP debug CLI & code uses rx/tx when refering
to packets comming into VPP (rx) or out of VPP (tx).

In Calio, the semantic is the opposite, the point of view taken by
the implementation is that of a container. Hence the traffic
coming out of VPP will be (ingress), and that coming into VPP
will be (egress).

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>