build: Fix api-proxy build and dev build.
This patch fixes the api-proxy build that could result in a mismatch between
the glibc version used for linking when building and the one used in the run
environment (i.e. the calico container where the image is run).

This also makes dev build produce :latest images in addition to the new
:commit-hash tags, to maintain old behavior.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>