agent: fix network watcher
Prior to this patch we were not correctly restarting the api watcher
for the network objects, ie when the watcher breaks network server in
calicovpp agent logs warning or returns an error and we only recover if
we restart agent. This patch allows to restart the watcher if it breaks,
while keeping the resource version to continue watching.