simplify punt fake gateway
This patch simplifies the way we handle the punt routes
when routing for-me traffic VPP cannot handle to the host.
Previously, we were using different fake gateways for ip4 & ip6.
We now use the ip4 indirection in both cases.
We also generate different hardware addresses using the uplink order
in the configuration.

Finally, we also make the isMain parameter public in the config.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>