add redirect to host rules for dns issue in kind
A dns issue showed in kind cluster, due to specific iptables rules added by docker.
This commit addresses that by adding a feature allowing to redirect some traffic to host,
to get it pass through iptables.
This is applied to coredns pods.
We use some new user config to specify rulesfor this feature.
A tag "host" is added to cnat snat poliy feature and we use it for our vpptap0 to
disable snat on traffic outgoing through it.