openshift: dhcp lease timeout issue
NetworkManager makes no attempt to renew dhcp lease when the default dhcp
lease time of 60 mins expires resulting in the wipe out of the node's ip
address. This fix coerces NM to keep trying to renew the lease
indefinitely so that node always has its ip address intact as long as
vpp is running.