cni: move the PBL to punt when VCL is on
This patch moves the PBL to the pod's VRF as it was previously
in the main VRF. This will enable pod to self communication
respecting the port mapping defined in the annotation.

When VCL is enabled on the pod, we add the PBL in the punt table
so that traffic is first looked up in VPP's session table, then
hits the port lookup in the PBL. This allows using all three
(memif,VCL,tuntap) for a single pod.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>