Support setting failsafepolicy net
This patch adds support for setting the .net property when defining
the failsafepolicies in felixconfig. This is a not so well documented
feature that slipped through when aiming for feature parity.

For reference, here is the documentation for [0] felixconfig.
- setting failsafeInboundHostPorts to e.g. {Net: "10.0.0.0/24", ...}
will add a filter for inbound packets for their source prefix to match 10.0.0/24.
- setting failsafeOutboundHostPorts to e.g. {Net: "20.0.0.0/24". ...}
will add a filter for outbound packets for their destination prefix to match 20.0.0/24

[0] https://docs.tigera.io/calico/latest/reference/resources/felixconfig#spec

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>