Disable memif if queue spread by default
This patch disables memif interface queue spreading on workers
by default. Users can enable it again using the debug knob :
CALICOVPP_DEBUG='{"spreadTxQueuesOnWorkers": true}'.

This is a consequence of the pinning behaviour of VPP by default.
When #queues < #workers, queues are shared by default and pinning
them removes their sharing rendering them unusable on some workers.

When #queues > #workers, the extra queues are unpined by default,
rendering multi-tx harder to use.

As the latter feature is harder and more niche than the former,
we chose to go with a default that suits most usecase, while we
work on a better pinning framework.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>