use single pbl index per pod
This patch addresses an issue where upon dataplane restart
a pod with PBL enabled would find itself with multiple PBL
instances as we did use a list and did not check for existing
entries on recreation.
This patch solves this using a single entry per pod while
waiting for a proper refactoring.