Add auto corepining path
This patch adds support in VPP for pinning workers to the CPU
cores that are assigned by a static allocator.

When k8s is configured to run with a static allocator [0]
and that VPP starts with requests.CPU equal to limits.CPU
and both integer numbers.

having VPP configured with

cpu {
   main-core 0
   corelist-workers 2-3
   relative
}

Will make it so that numbers 0,2,3 do not refer to absolutes CPU
IDs but to the 0th, 2nd and 3rd CPU of those allotted by the static
core allocator in k8s.

[0] https://kubernetes.io/docs/tasks/administer-cluster/cpu-management-policies/#static-policy

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>