fix auto hep behavior and policies ordering
A normal hep is expected to deny all traffic if no policies are specified
that is not the case for an auto hep, as it should allow all traffic using the default allow profile.
Here we remove the implicit deny all policy in the case where there are profiles applied.
We also give workloadToHostPolicy priority over failsafe rules to comply with calico.