Fix makefile VPP_HASH cache invalidation
This patch addresses the fact that the VPP_HASH computation
did not contain the list of deb files included within the
VPP build cache. As a consequence, expanding this list, e.g.
when added extra deps breaks the build.

This patch addresses this by making the deb list being part of
the VPP_HASH computation.

This also enables the tar.gz computation when running locally, that way
if the vpp_clone_current.sh has not changed, we won't rebuild VPP

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>