change capo to make default behavior configurable
This patch uses a vpp capo plugin patch that makes the default behavior
when policies don't match configurable: drop, or goto profiles, depending
on the existence of user defined policies
This addresses the issue where dataplane takes wrong decisions because it
isn't aware of the difference between a user defined policy and an implicit
policy (like failsafe)
This also prepends an explicit deny policy in the case of an empty hostendpoint
(hep with 0 policies and 0 profiles).
This also makes endpointtohost action prior to failsafe to comply with doc.