Split CNI into watcher/handler under felix
This patch splits the CNI watcher and handlers
in two pieces. The handling will be done in the main
'felix' goroutine, while the watching / grpc server
will live under watchers/ and not store or access agent
state.

The intent is to move away from a model with multiple servers
replicating state and communicating over a pubsub. This being
prone to race conditions, deadlocks, and not providing many
benefits as scale & asynchronicity will not be a constraint
on nodes with relatively small number of pods (~100) as is k8s
default.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>