Split Connectivity into watcher/handler under felix
This patch moves the Connectivity handlers in the main felix
loop to allow lockless access to the cache.

The intent is to move away from a model with multiple servers
replicating state and communicating over a pubsub. This being
prone to race conditions, deadlocks, and not providing many
benefits as scale & asynchronicity will not be a constraint
on nodes with relatively small number of pods (~100) as is k8s
default.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>