Split Services into watcher/handler under felix
This patch splits services in two components,
- a watcher that handles the informer fetching services and
endpoints from the k8s API.
- a handler that takes care of programming VPP with the NAT
rules, within the context of the felix server's single goroutine.

The intent is to move away from a model with multiple servers
replicating state and communicating over a pubsub. This being
prone to race conditions, deadlocks, and not providing many
benefits as scale & asynchronicity will not be a constraint
on nodes with relatively small number of pods (~100) as is k8s
default.

Also cleaned up unused code from single-thread agent refactor.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>