yaml: fix eks breakage in AL2023
Amazon Linux 2 (AL2) has been deprecated and in its place has come
Amazon Linux 2023 (AL2023). AL2023 has many improvements and many more
changes and sadly it breaks the deployment of Calico/VPP in eks. Calico/VPP
causes the uplink interface to vanish from the root network namespace which
confuses systemd-networkd and it reacts by wiping out the DNS config and refusing
to configure/manage the tuntap interface created in place of the
original uplink interface which in turn causes the kernel to purge the
tuntap interface at the expiry of the 1 hour DHCP lease.