Move to json storage, address pblindex issue (#721)
This patch changes the way we persist the data on disk when
running Calico/VPP. Instead of using struc and binary format
we transition to json files. Size should not be an issue as number
of pods per node are typically low (~100). This will make
troubleshooting easier and errors clearer when parsing fails.

We thus remove the /bin/debug troubleshooting utility as the
data format is not human readable.

Doing this, we address an issue where PBL indexes were reused
upon dataplane restart, as they were stored in a list. We now
will use a map to retain the containerIP mapping.

We also split the configuration from runtime spec in LocalPodSpec
and add a step to clear it when corresponding VRFs are not found
in VPP.

Finally we address an issue where uRPF was not properly set up
for ipv6.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>