Fix the weird DHCP/DNS issue
When Calico/VPP is deployed, one of the things that happens is that the
uplink interface vanishes from the root network namespace. Depending on the
uplink driver used, it is either moved to a different network namespace or it
vanishes completely from the kernel/os, and it gets replaced with a tap
interface. This might cause an issue on systems where instead of the usual mac
address or the interface name, systemd-networkd's [Match] section is configured
with some parameter like the "permanent mac address" which is either not
supported or not set for tap interfaces, thus, causing systemd-networkd to not
manage the tap interface. This has the undesirable effect of the dhcp
lease expiring and not getting renewed and bricking the system, and in
additon also wiping out the DNS configuration.

With this fix, the [Match] section of the network file associated with the
uplink interface is tweaked to use "Name=<interface name>".