Harmonize makefile docker build proxy and fix multiplatform
This patch harmonizes the proxy settings to be used for build.
We pass --network=host and allow setting http_proxy and https_proxy
to a provided value.

We also prevent the build to fail is the containerd-snapshotter is
not enabled on the docker daemon where we run.
This will output a warning and instructions to enable but will not
cause the build to stop.

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>